fix: Increased the delay before initial puzzle sequence displays
Added a variable that can be adjusted to allow more time before the
puzzle starts the first time after clicking the start button, to
account for the slower travel speed to reduce motion sickness.  variable
can then be reset to shorter interval if user fails and pattern needs
to be displayed again.